\docType{package}
\name{cleaver-package}
\alias{cleaver-package}
\title{Cleavage of polypeptide sequences}
\description{
  This package cleaves polypeptide sequences. It provides
  three functions: \code{\link[cleaver:cleave-methods]{cleave}},
  \code{\link[cleaver:cleave-methods]{cleavageRanges}} and
  \code{\link[cleaver:cleave-methods]{cleavageSites}}.

}
\details{
  The cleavage rules are taken from:
  \url{https://web.expasy.org/peptide_cutter/peptidecutter_enzymes.html}

  \tabular{ll}{
    Package: \tab cleaver\cr
    License: \tab GPL (>= 3)\cr
    URL: \tab https://codeberg.org/sgibb/cleaver/\cr
  }
}
\author{
  Sebastian Gibb \email{mail@sebastiangibb.de}
}
\references{
  \url{https://codeberg.org/sgibb/cleaver/}

  Gasteiger E., Hoogland C., Gattiker A., Duvaud S.,
  Wilkins M.R., Appel R.D., Bairoch A.; "Protein
  Identification and Analysis Tools on the ExPASy Server".
  (In) John M. Walker (ed): The Proteomics Protocols
  Handbook, Humana Press (2005).\cr
  \url{https://web.expasy.org/peptide_cutter/peptidecutter_enzymes.html}
}
\seealso{
  \code{\link[cleaver:cleave-methods]{cleave}},
  \code{\link[cleaver:cleave-methods]{cleavageRanges}} and
  \code{\link[cleaver:cleave-methods]{cleavageSites}}.
}
\keyword{package}