\docType{package}
\name{cleaver-package}
\alias{cleaver-package}
\title{Cleavage of polypeptide sequences}
\description{
  This package cleaves polpeptide sequences. The cleavage
  rules are taken from:
  \url{http://web.expasy.org/peptide_cutter/peptidecutter_enzymes.html}
}
\details{
  \tabular{ll}{ Package: \tab cleaver\cr Version: \tab
  0.1\cr Date: \tab 2013-04-07\cr License: \tab GPL (>=
  3)\cr URL: \tab https://github.com/sgibb/cleaver/\cr }
}
\author{
  Sebastian Gibb \email{mail@sebastiangibb.de}
}
\references{
  \url{https://github.com/sgibb/cleaver/}
}
\seealso{
  \code{\link[cleaver]{cleave}}
}
\keyword{package}

