\docType{package}
\name{cleaver-package}
\alias{cleaver-package}
\title{Cleavage of polypeptide sequences}
\description{
  This package cleaves polypeptide sequences. It provides
  one function: \code{\link[cleaver]{cleave}}
}
\details{
  The cleavage rules are taken from:
  \url{http://web.expasy.org/peptide_cutter/peptidecutter_enzymes.html}

  \tabular{ll}{ Package: \tab cleaver\cr Version: \tab
  0.7\cr Date: \tab 2013-04-10\cr License: \tab GPL (>=
  3)\cr URL: \tab https://github.com/sgibb/cleaver/\cr }
}
\author{
  Sebastian Gibb \email{mail@sebastiangibb.de}
}
\references{
  \url{https://github.com/sgibb/cleaver/}

  \url{http://web.expasy.org/peptide_cutter/peptidecutter_enzymes.html}
}
\seealso{
  \code{\link[cleaver]{cleave}}
}
\keyword{package}

