\docType{package}
\name{cleaver-package}
\alias{cleaver-package}
\title{Cleavage of polypeptide sequences}
\description{
  This package cleaves polypeptide sequences. It provides
  only one function: \code{\link[cleaver]{cleave}}
}
\details{
  The cleavage rules are taken from:
  \url{http://web.expasy.org/peptide_cutter/peptidecutter_enzymes.html}

  \tabular{ll}{
    Package: \tab cleaver\cr
    Version: \tab 0.99.4\cr
    License: \tab GPL (>= 3)\cr
    URL: \tab https://github.com/sgibb/cleaver/\cr
  }
}
\author{
  Sebastian Gibb \email{mail@sebastiangibb.de}
}
\references{
  \url{https://github.com/sgibb/cleaver/}

  Gasteiger E., Hoogland C., Gattiker A., Duvaud S.,
  Wilkins M.R., Appel R.D., Bairoch A.; "Protein
  Identification and Analysis Tools on the ExPASy Server".
  (In) John M. Walker (ed): The Proteomics Protocols
  Handbook, Humana Press (2005).\cr
  \url{http://web.expasy.org/peptide_cutter/peptidecutter_enzymes.html}
}
\seealso{
  \code{\link[cleaver]{cleave}}
}
\keyword{package}

